Use consistent naming for China
The recipient country codelist is lifted from IATI codelists, and is the source for the select boxes used on the activity form to select the recipient countries.

However, in this case, we don’t want to create a BEIS-specific codelist file, and accept that if we ever replace this file with a new codelist release from IATI, we might have to do the change again.

The IATI locale file was compiled by us based on official IATI codelists, and is used to display country names in both recipient country and intended beneficiaries sections on the activity details page.