Send report approved email notification to BEIS users
Previously, only delivery partner users belonging to the organisation whose report has been approved would be notified by email.

Some BEIS users want to receive notifications when any report is approved.

We don’t have a way to selectively send only to some BEIS users but not others, so for now we’ll send it to all of them and iterate if needed.