Service for exporting all the transactions
Given a set of activities, QuarterlyTransactionExport generates a
CSV-compatible response that shows the total actual spend by quarter for
all the given activities.

All the transactions linked to the given activities are loaded in a
single query, and their earliest and latest financial quarter is found.
We export a column for every quarter in this range, putting "0.00" in
cells where no spend is recorded for that activity and quarter.

So the columns in the response are the activity RODA and BEIS
identifiers, followed by every quarter from the earliest to the latest
we have records for. The rows are all the activities in the set,
including ones with no spending recorded.

This design lets us export a complete set of all data across all
organisations for a BEIS user, or just the activities for a particular
organisation for a delivery partner.