Move Aid type question
We will soon be adding logic for questions who's responses can be
calculated based on the answer to this question.

One of those dependant questions is Collaboration type, so the aid type
question must come before it.

As a first step to enabling that work, we move the Aid type question up
the order to before the Collaboration type question.