Stop using the report description as the title
By not using the report description in the report view page title we can
make it optional.

We have learnt that until we see two reports in the same financial
quarter, the description has little value. So far all report
descriptions have been the same <financial quarter><fund> which is
something the application can do itself and which makes the title feel
duplicated.

To enable the report description to be optional we stop using it in the
report page title.