Extend the report summary
We already showed the report state as summary on a single report.

Here, we add further helpful information for users:

- the report deadline
- the organisation of the report (if a BEIS user)
- the description of the report, if there is one
- whether the user can change (edit) the report or not

This last one has been added as we think some users may not understand
the relationship between report state and if the report contents can be
changed. We feel this is a nice, simple, unobtrusive way to show that.