Update the FSTC applies template
This updates the fstc_applies template to remove the
wording that prompts the user to select an option based
on their answer to the aid type question. It also adds the
code in brackets to be consistent with how we prompt
users in other questions.

I also noticed that, because we were specifying the
radio buttons manually, rather than using a `collection`
method, the link to the help documentation wasn't
showing, so I've fixed that too.