Create a BEIS codelist for collaboration type
In RODA we don't want to use the full list of collaboration types
defined by IATI, we only want to allow:

- Bilateral
- Multilateral (inflows)
- Bilateral, core contributions to NGOs and other private bodies / PPPs

This adds a BEIS codelist that's a copy of the IATI one, only including
the allowed values. None of the activities in production have a
collaboration type that's not one of the above.