Fix transaction export spec array assertion
This test always fails for me, but it doesn't appear to fail on CI or
on other people's machines. It could be because I'm running a newer
version of Postgres (13.2) -- in any case we don't specify the return
order for the transaction data, so we can use `match_array` instead.