Add a `reportable` spec to activities
This adds a new scope to the Activity model to return all the activities that are considered
"reportable". This is activities that have an ODA Status of "eligible" and are do not have a
programme status of "completed", "stopped", "cancelled" or "paused".

We already have  a `current` scope, whch checks for activities that are not "completed",
"stopped" or  "cancelled", which we could chain on, but I felt it was clearer to have a
`NON_CURRENT_PROGRAMME_STATUSES` constant, which we add "paused" to to make
a `UNREPORTABLE_PROGRAMME_STATUSES` constant.

We could also have used the enum scope `eligible` to return all ODA eligible activities, but
I felt it was clearer to query to column directly as just calling `eligible` doesn't make it
clear _what_ column we're querying.