Replace aid type B02 with D01 in tests
We're about to add some logic to infer the collaboration type if the aid
type is either B02 or B03. This would cause any tests using the B02
value to display non-default behaviour, so causes various tests to
break.

I'm changing this to D02, a value that's quite common in production and
which doesn't cause any other fields to be automatically inferred.