Prevent changing the DP identifier through upload
We’ve decided to reject an upload in the same way we reject it for RODA identifiers that would be overwritten.

The assumption is that an existing activity will have a DP identifier already set (it would be invalid otherwise), so it’s sufficient to check whether we’re in the `create` or `update`.