Set the collaboration type based on the aid type
We need to implement the following rules for the value of the aid type:

- If it is B02, then collaboration must be 2: Multilateral
- If it is B03, then collaboration must be 1: Bilateral

If aid type is any other value then collaboration type is a free choice
and left up to the user.

We enforce this by making Activity::Updater set the value of
`collaboration_type` if `aid_type` is set to one of the above values,
and making ActivityFormsController skip the `collaboration_type` step in
this case.