Introduce a service for inferring field values
We have a growing set of scenations where setting the value of one
Activity field determines the value of another. For example setting
`aid_type` to certain values currently determines the value for
`collaboration_type` and `fstc_applies`.

As this set of rules grows it will be more of a liability to have their
logic scattered across different parts of the codebase. Currently we
need to make at least the following changes when one of these rules
comes in:

- Update `Activity::Updater` to update a dependent field when a source
  field (e.g. `aid_type`) is set

- Update `Staff::ActivityFormsController` to skip the dependent field's
  step if its value is fixed by the source field's value

- Update views to hide any links to the dependent field's step if it
  cannot be edited

As the set of rules grows we will have situations where changing field
`a` causes a change to field `b`, which in turn causes a change to field
`c` and so on, and maintaining those effects when they're split across
various form steps is likely to introduce bugs.

To make this more maintainable, I'm introducing a service named
FieldInference that can encode and execute rules for changing dependent
fields. In this first commit it just supports a single kind of rule
where setting one field can change the value of another.

For example, we can write a rule that setting field `x` to value 1 fixes
the value of `y` to 2.

    rules = FieldInference.new
    rules.on(:x, 1).fix(:y, 2)

The word "fix" indicates that we're not just setting the value of `y` at
this moment, but making it non-editable so that whenever `x` is 1, `y`
cannot be changed.

We execute these rules by using this service to change an object's field
values, for example calling `assign(point, :x, 1)` also changes the
value of `point.y`:

    Point = Struct.new(:x, :y)

    point = Point.new(0, 0)
    rules.assign(point, :x, 1)

    # => point = #<struct Point x=1, y=2>

This also causes `y` to become non-editable, which we enforce by calling
`editable?(point, :y)`:

    rules.editable?(point, :y) # => false

If the value of `x` is changed, `y` becomes editable again -- the
editability of a field is based on the value of the object's properties,
not by storing any additional flags about fields being locked.

    point.x = 2
    rules.editable?(point, :y) # => true

We'd use the `editable?` method to skip steps and hide links to fields
when they cannot be freely modified by the user.