Handle cascading dependent values in FieldInference
When editing activities, we will sometimes have rules where changing
field `a` causes a change to field `b`, and changing field `b` causes a
change to field `c`. We can use FieldInference to encode these rules so
that these cascading changes are automatically applied, and we don't
need to modify the code for steps `a` and `b` to modify `c` explicitly.

Here's an example where setting `x` to 1 fixes `y` to 2, and setting `y`
to 2 fixes `z` to 3.

    rules = FieldInference.new
    rules.on(:x, 1).fix(:y, 2)
    rules.on(:y, 2).fix(:z, 3)

Now we can call `assign(point, :x, 1)` and all the changes that result
from this will be applied.

    Point = Struct.new(:x, :y, :z)

    point = Point.new(0, 0, 0)
    rules.assign(point, :x, 1)

    # => point = #<struct Point x=1, y=2, z=3>

This will also make `z` non-editable:

    rules.editable?(point, :z) # => false

If we change `x`, `z` is still non-editable because it's fixed by the
fact that `y` is 2.

    point.x = 2
    rules.editable?(point, :z) # => false

But if we change `y`, then `z` becomes editable again:

    point.y = 0
    rules.editable?(point, :z) # => true