Set a `target_activity` in the transfer controller
For Outgoing transfers, we want to set the report for the source activity
when creating a transfer, as well as redirecting to the source activity's path
when creating/editing. For incoming transfers, we want this to apply to the
destination activity, so we set a `target_activity` method that returns the
source_activity. In the Incoming transfer, we'll set this to be the
`destination_activity`