Handle conflicting requirements in FieldInference
With dependent field logic scattered across controllers, it is easy to
introduce bugs where two fields each trigger different values in a third
field. Here's an example where setting `x` to 1 fixes `z` to 3, but
setting `y` to 2 fixes `z` to 4:

    rules = FieldInference.new
    rules.on(:x, 1).fix(:z, 3)
    rules.on(:y, 2).fix(:z, 4)

If this logic is hand-coded into the form steps for `x` and `y`, it's
easy for those steps to each set `z` to a different value and for nobody
to notice these fields are having contradictory effects.

With FieldInference we can detect these cases and get an exception when
two rules produce a conflicting requirement. In this example, setting
`x` to 1 is fine, it just changes `z`:

    Point = Struct.new(:x, :y, :z)

    point = Point.new(0, 0, 0)
    rules.assign(point, :x, 1)

    # => point = #<struct Point x=1, y=0, z=3>

But if we then try to change `y`, we get an exception:

    rules.assign(point, :y, 2)

    # => Cannot set `y` to 2: would change the value of `z` which is
         fixed to 3 because `x` is 1 (FieldInference::Conflict)

Whenever `assign` is called, we check that there are no rules
activitated by the object's current field values that cause a dependent
field to be fixed to multiple different values, and the assignment only
works if no conflicts are detected.