Use FieldInference to restrict fields to a set of options
So far FieldInference has let us define rules of the form:

    when field `a` is set to value `x`
    then fix field `b` to value `y`

We also need to write rules of the form:

    when field `a` is set to value `x`
    then field `b` may have value [`y`, `z`, ...]

That is, we don't want to set `b` immediately, it remains editable, but
its allowed values are restricted by the value of `a`.

Here's an example saying that if `x` is 1, then `y` is allowed to be 2,
3 or 4.

    rules = FieldInference.new
    rules.on(:x, 1).restrict(:y, [2, 3, 4])

When `x` is not set to a matching value, `y` is editable and
`allowed_values` returns `nil` indicating there's no restriction on this
field and we can pick whatever value for it we like.

    Point = Struct.new(:x, :y)

    point = Point.new(0, 0)
    rules.editable?(point, :y) # => true
    rules.allowed_values(point, :y) # => nil

When `x` is set to 1, `y` remains editable and `allowed_values` now
returns a list of the valid values for `y`:

    rules.assign(point, :x, 1)
    rules.editable?(point, :y) # => true
    rules.allowed_values(point, :y) # => [2, 3, 4]