Handle edge cases when retricting the values of fields
When we have rules where the value of field `a` determines the allowable
values for field `b`, it's possible for these rules to exhibit a couple
of edge cases:

- Single-value restriction: two active rules restrict the values of `b`
  to different sets, with a single element in common. In this case the
  value of `b` should be fixed to this element and `b` should become
  non-editable.

- Empty-set restriction: two active rules restrict the values of `b` to
  completely different sets with no common elements. This makes it
  impossible to select a valid value for `b` and this should be treated
  as an error.

Here's an example. Setting `x` to 1 restricts `z` to 3 or 4, and setting
`y` to 2 restricts `z` to 4 or 5.

    rules = FieldInference.new
    rules.on(:x, 1).restrict(:z, [3, 4])
    rules.on(:y, 2).restrict(:z, [4, 5])

When we assign `x` to 1, that doesn't by itself change the value of `z`.
`z` remains editable but has a restricted set of allowed values.

    Point = Struct.new(:x, :y, :z)

    point = Point.new(0, 0, 0)

    rules.assign(point, :x, 1)
    # => point = #<struct Point x=1, y=0, z=0>

    rules.allowed_values(point, :z) # => [3, 4]
    rules.editable?(point, :z) # => true

If we then set `y`, the combined rules for `x` and `y` give only one
possible value for `z`. So, `z` becomes fixed to that value:

    rules.assign(point, :y, 2)
    # => point = #<struct Point x=1, x=2, z=4>

    rules.allowed_values(point, :z) # => [4]
    rules.editable?(point, :z) # => false

In general, a field is editable if it has more than one allowed value,
or is unrestricted. If this is not the case, the field's value is
assigned and it becomes non-editable.

These changes also cause FieldInference to raise an exception if an
editable field is assigned a value outside its allowed set:

    rules.assign(point, :z, 7)
    # => Cannot set `z` to 7: would change the value of `z` which is
         restricted to [3, 4] because `x` is 1