Migrate existing dependent field logic to FieldInference
We currently have rules about the `aid_type` field. When it is set to
certain values, we can infer the values of `collaboration_type` and
`ftsc_applies` from it. This logic is currently implemented in the
individual step-handling logic in Staff::ActivityFormsController and
Activity::Updater using the `can_infer_collaboration_type`,
`can_infer_fstc`, `collaboration_type_from_aid_type` and
`fstc_from_aid_type` helper functions.

Here we replace those implementations with the use of a FieldInference
instance populated from the `aid_type` codelist, and accessible as
`Activity::Inference.service`. This updates any dependent fields as
required and deals with skipping steps and hiding links when fields are
non-editable.