Rename planned disbursements to forecasts
The duplicate terminology is creating confusion. “Planned disbursement” originates from IATI, but the users (mainly BEIS and delivery partners) think of them as “forecasts”.

A lot of the initial development of RODA was based on IATI reporting needs, which is how we ended up with IATI terminology.

We want to prioritise the users’ domain language, and we value consistency to make it easier for the delivery team to discuss these concepts, so we’ve decided to use “forecasts” everywhere except within the XML files generated for IATI reporting.