Add a script to prepare for an import in dev
As part of the import process for newly onboarded DPs, we need to test the
data in development.

As part of this process, to make sure our local  environment mirrors that of
the production environment, we need to take  a cut of the production data
and import it to our local environment.

As well as this, we need to make sure  that we have a user for the DP we are
testing with, as well as an active  report. As we have to do this a number of
times for each DP, this needs to be done a lot, so I've added a script that
allows us to automate this, to make the testing process less painful.