Set inferred fields to nil if their current value is not allowed
Say we have the following rules where setting `x` to 1 restricts the
value of `y` to 2 or 3:

    rules = FieldInference.new
    rules.on(:x, 1).restrict(:y, [2, 3])

If we have an object with `y` equal to 1:

    Point = Struct.new(:x, :y)
    point = Point.new(0, 1)

Then setting `x` to 1 means the current value of `y` is not longer valid
-- if `x` is 1 then `y` must be 2 or 3. Therefore we remove the current
value of `y`:

    rules.assign(point, :x, 1)
    # => point = #<struct Point x=1, y=nil>

This design is intended to address an upcoming requirement where setting
`collaboration_type` to 3 on an activity limits the possible choices of
`channel_of_delivery_code` to a setset of the allowed values. It's legal
for most Activity fields to be nil, because we have a multi-step form
for incrementally setting the fields.

We should arrange these steps so that a field always comes later than
any fields that could determine its value, so that users cannot set a
value and then have it implicitly removed by later steps.