Add table for forecasts grouped by activity
The table currently includes a row for each activity and
the:

- activity title
- activity RODA ID

We include an 'activity id' in the markup on each major row.
This allows us to describe the way in which activity (and
its forecasts) are represented in the view -- from our
tests.

I'm comfortable adding this information to the markup,
initially as a facility to drive the implementation from our
acceptance test because:

1. it makes for robust and maintainable tests as each
   activity's expected representation is defined with

   expect_to_see_a_table_of_forecasts_grouped_by_activity(activities)

2. I anticipate that when  we come to describe how the
   'upload a forecast'feature redirects to this new 'Forecasts'
   tab, it's going to be useful to be able to link to a
   specific activity so that the user can see the newly
   uploaded forecast in the context of the activity to which it
   belongs.