Add guards to verify presence of meaningful fixtures
Our `expect_to_see_a_table_of_forecasts_grouped_by_activity`
assertion assumes that activites will be present and will
have forecasts to display. But if something goes wrong with
the fixtures and the expect activites and forecasts were
missing then we could end up asserting nothing.

We add a couple of guard clauses to raise the alert in that
eventuality.