Add a data migration to set budget types correctly
All the Budgets we have in prod at the moment are `1` or `2` (so all direct),
so we need to add a data migration to set this to `0` (which is how direct
budgets are recorded in our enum column now).

Because enums are now used, running `Budget.where(budget_type: [1,2])`
won't work, because `2` is no longer a valid budget_type, so we have to
build the query (semi) manually.