Render the existing policy marker values on edit
Any existing values for the `policy_marker_*` fields were not recognised when rendering the radio buttons, and the radio buttons were always rendered as if there were no values selected. This had the effect of overwriting those values with the `not_assessed` default.

This was due to a mismatch between their value as an integer, and the corresponding string (e.g. `1` versus `significant_objective`).

Being enums, the value stored in a database is an integer, but the value of the field from Rails’ perspective is a string, so we need to make the value of the radio button a string so that the Rails helpers recognise it as selected.