Show a meaningful error message for invalid policy markers
When importing an activity from a CSV file, users are expected to provide the numeric codes for the policy markers. (This is in contrast to the activity form, where the selection is handling any code to string translation and viceversa behind the scene.)

Previously there was no error message to let the user know if they had used an invalid code. Instead, they would have been shown the misleading message that the policy marker “can’t be blank”. (Which is also true [1], but didn’t reflect that the blankness came from the code not being recognised by the importer.)

[1] If the column is left blank, the importer fills in the default `not_assessed` code.