Make Budget Type an enum
Now we've done away with a lot of budget types, it makes sense for us to
simplify how we set them. Rather than having a codelist of budget types, we
have an `enum` of budget types - either "direct" or "other_official". Because
we know the fund from the parent activity, there's no need to specify if the
"direct" type is either Newton or GCRF