Handle variety of data types, using AR serialisation
see https://apidock.com/rails/v6.1.3.1/ActiveRecord/AttributeMethods/Serialization/ClassMethods/serialize

which stores the value as YAML in the db.

The two "value fields":

- `previous_value`, and
- `new_value`

need to handle a variety of data types. At present this is
limited to:

- `string`
- `integer`
- `date`
- `datetime`
- `boolean `

which are all data types currently in use on attributes of
`Activity`.