Fix tab navigation aria elements
The tab item elements all has an `aria-contols` value of
`financials` and a `selected` value of `false`. Aria tab
roles and their associated attributes should only be used
if we are using a JS-powered tab navigation where all the
elements are present on the page, but the tabs control
their visiblity.

The way we use tabs here is closer to a navigation list, so
I’ve added a `role` of `navigation` to the tablist, as well
as setting `aria-current` to `page` to signal to the user
which is the current page.

More info on `aria-current` is here:

https://www.aditus.io/aria/aria-current/