Introduce a basic home page
This commit lays the ground work for a real home page that users see
upon signing in.

This home page will vary its content based on which organisation the user
is associated with.

Here we introduce the route, the view, the policy and the controller.

The policy checks the user is either a BEIS or delivery partner.

As this work only includes BEIS users, other users are redirected to
their organisation page, which is the current behaviour.

The route is not yet used so has no impact (yet!) but can be reached at
`/home` when signed in.

This commit also removes some very old 'dashboard' code that probably
served a similar purpose and has never been removed.