Show table of delivery partners on BEIS home page
This lets BEIS users choose the delivery partner they want to view right
on the home page without having to use the organisation section.

Note: the back link on the organisation show page is broken by this
change in that it is not link 'back' as there are multiple placed from
which you might have come the page we will be looking to fix this as
part of wider work centred on breadcrumbs.