Verify that report is associated with HistoricalEvent via integration test
Note that the "as a BEIS user" scenario has no report
associated with the activity being edited.

Either:

a) this describes a valid use-case when an activity which
has no associated report is edited -- and is a good example
for why HistoricalEvent::belongs_to Report is optional, or

b) there's a deficiency in the both i) the setup of this
test and ii) the controls in place which ought to be
preventing an Activity which has no report from being
edited, or

c) there more to this than I understand. Quite likely!