Use the new home page for BEIS users
All 'home' links ad the sign in redirect lead to the home page.

This leaves delivery partners with a odd redirect from `/home` to their
organisation show page, this will be resolved in our next piece of work
where we use the home page for delivery partners as well as BEIS users.