Return nil in presenter if budget_type is nil
At the moment, if the budget_type is nil, the presenter
outputs a hash of all the available `budget_type`s from
the translations. Even though we treat a Budget without
a `budget_type` as invalid now, there still will be the odd
one kicking around, so we should fix this. It’s also more
consistent with how other presentes handle things.