Show the hierarchical list of activities
On the delivery partner home page we show the hierarchical list of
activities so that a user can quickly browse to an activity record.

We've reused the existing partial for this which means duplicating the
`fetch_grouped_programmes_for` method in the home controller. It feels
like this method might be better served as a scope on the Activity
model, but we decided to keep the duplication for now as the method also
results in some unexpected child activities are shown (those in
incorrect statuses) so we will look to fix those issues together as
future work.