Fix: vistors controller redirect
Whist working on the Delivery partner home page I noticed this
`visitors` route is also the `root_path` and that it has it's own
redirect and test, this have been updated so the redirect is to the home
page and not the old organisation show page.

`root_path` seems to mainly be used in tests and error pages which is
how I think it was missed.