Activities on homepage are shown in tabs
Delivery partners see a list of activities split by fund on their
home page.

We have used the GOVUK tab component to progressively enhance these two
stacked lists into tabs for users that have javascript available.

As part of this work we couldn't see the GOVUK JS being initialised so
we corrected that.

We wanted to keep the hierarchical activity view at `/activities/`
tab-less so we duplicated the `tree_view` partial and added the required
markup - this felt simpler than attempting to keep a single partial for
both types.