Report shows updated activities
The goal of the report activities tab is to show all users what new and
updated activities are in a report.

Building on the new method on Report `activities_updated` and
`new_activities` association, we can now do this.

Here we populate the existing table with updated activities only, changing
the table elements and caption to reflect this.

We also moved the upload section of the tab under a heading and
collected the guidance and upload button together under the heading to
visually break up the sections.