Extract grouping code to HistoricalEventGrouper
This is a preliminary to facilitate the next step:
describing how the 'grouper' orders the groups of historical
events, newest first.

The unit test for `HistoricalEventsGrouper` was causing 'Bullet'
(https://github.com/flyerhzm/bullet) to wrongly complain
that the eager-loading of associated 'reports' was not
necessary. Hence the addition of:

    Bullet.add_whitelist
      type: :unused_eager_loading,
      class_name: "HistoricalEvent",
      association: :report

to the test env config.