Order events, newest first
By extracting the finding, ordering and grouping of
historical events into a service class we are able to refine
its behaviour in a unit test, rather than by creating
additional slow integration tests.

Regarding the ordering which is introduced in this commit,
it's worth noting that `Enumerable#group_by` returns an
`OrderedHash`. At first glance it looks like a regular Ruby
`Hash`, in which ordering is not reliable, but in fact it is
an `OrderedHash`. See:

https://apidock.com/rails/Enumerable/group_by#820-returns-an-ActiveSupport-OrderedHash

Note: if the quantity of historical events associated with
an individual activity becomes very large, then we might see
some performance/memory issues here, so something to keep an
eye on...