Use activities fetcher in controllers
This changes the controllers to use the new grouped
activities fetcher when displaying the activities in the
table with drop down displays of projects and
programmes.

I’ve had to use `skip_policy_scope` in the
ActivitiesController as the BaseController we inherit from
calls `verify_policy_scoped` after each index action. Even
though we’re using a scope in the fetcher,
`verify_policy_scope` doesn’t pick this up, as we dont
use the `policy_scope` helper, which sets a
`@_pundit_policy_scoped` instance variable, which
`verify_policy_scope` uses to check the scope has been
used.

See https://github.com/varvet/pundit#ensuring-policies-and-scopes-are-used for more detail.