Update ChangeLog
So that users, when viewing a report, are able to
distinguish between:

- "new" activities: which were created during this reporting
  period, and

- "updated" activities: updated during the period but created
  during a previous period

We need a clear association between an activity and the
report which was current when it was made. This was not
achieved with the previous Report#activities_created method:

  def activities_created
    reportable_activities.where(created_at: financial_period)
  end

because I may create an activity on July 1st for a Q1
(Apr 1 to June 30) Report.

Trello references:

- Split activity report by 'new' and 'updated' activities:
  https://trello.com/c/DCjBwQO0/1948-split-activity-report-by-new-and-updated-activities

- Associate activity with "originating" report:
  https://trello.com/c/ib0L8Qdy/1994-the-way-we-calculate-reportactivitiescreated-is-incorrect