Fix activities_updated
This methods gets all a Report historical events and fetches all the
associated activities.

In rare cases the activity which the historical event relates has been
deleted and so the `find` call raises an error.

We can use a `where` to guard against this and keep the application from
raising an error, but we do need to resolve the underlying issue of
orphaned historical events and why they occur in the first place.