Destroy HistoricalEvents when deleting Activities
There currently no value in storing historical events where the parent
activity has been deleted. There is no way to display the events or
utilise them.

If an activity has been identified to be deleted it usually means it
was made in error or it contains an error that cannot be rectified so
that the only answer is to delete it, we can see little value in keeping
the history of such an activity.

For these reasons we want a deleted activity to remove its historical
events.