BEIS users have a custom activities#index view
BEIS users want to see activities by delivery partner organisation.

We check the current users is a BEIS user (service owner) and if they
are render a list of delivery partners linking to the list of activities
filtered to the organisation.

The BEIS activities index page also includes a search widget to aid
locating a specific activity.

This change also means we only have to  execute the expensive
GroupedActivitiesFetcher when needed.

The `users_can_view_activites_spec` has been tweaked to reflect this
change and now assumes a BEIS users is already viewing a delivery
partners activities.

The old delivery partner filter is removed at this point.

This commit will break the test suite. We rely on a large number of
feature specs that navigate to an activity first before exercising it.

The result is that this commit needs a lot of small changes to the specs
to get the test suite green again.

I've decided to keep the commit that fixes this separate from this to
aid reviewing.

If we want to keep the test suite green we can squash this and the next
commit.