Set Transparency Identifier via Activity defaults
As we’re going to be removing the `cache_roda_identifier`
method in the next commit, we need a way to set the
transparency identifier when we create the roda_identifier.
There’s a couple of tests that we want to disable for now,
because the transparency identifier we expect isn’t the
one we get. This will change in the next commit.