Fix specs for change to activity navigation
In the last commit we added a step for navigating to activities, BEIS
users seeing a list of delivery partners rather than a drop down filter.

We rely heavily (too heavily!) on feature specs and a large number of
them follow a pattern that is something like:

- browse to an activity
- test some activity behaviour

The browse to an activity is not really under test, we have separate
tests for that behaviour, so can really be skipped.

One could argue that a good feature spec always includes all the steps a
user would take to perform the action and I agree, but counter that we
are over using feature specs in most cases here and the browse to is not
really bringing any value, feature specs are generally slow so anything
we can do to make them faster is a good thing.

With that in mind I have replaced all the cases where a failing feature
spec includes browsing to a activity, with a simple call to the activity
path itself, so now we have something like:

- load an activity view
- test some activity behaviour

I've also tried to:

- move any feature tests for a signed out user being redirected to
  controller tests
- tweak some words