Change activity routing
Now we no longer have the single view of activities with its drop down
filter, we want to change the routing to:

- show a delivery partner activities at `/organisation/id/activities`
- when beis view this route, the delivery partner list is shown, linking
  to individual delivery partner activities views
- support the 'current' and 'historic' activity split

We also tweak the activity index page to show the organisation name
which aids BEIS users at no real cost to delivery partners and show the
activity search widget.

With this in place the activity listing on the organisation show page
are no longer relevant and have been removed.