Send a notification on successful deploy
This adds a new job in the deploy Github Action to poll
the Health Check URL and check the SHA returned
matches the current SHA of the code we’ve deployed.
Once it does it sends a message to our Slack channel to
inform the team that the code has been deployed.

This only runs on production deploys and times out after
30 minutes (in case something has gone wrong and we
end up waiting forever for something that is never going
to happen)