Move new programme (level B) activities buttons
Previously these buttons where on the organisation show page. Now we
have an activities index page for each delivery partner, this is a more
natural home for this action.

Most of the changes here are to feature specs where we are navigating to
the old location in order to click the 'add new level B activity'
button, which has moved..

In principle we would want a BEIS users to be able to add a new
programme (level B) activity even if the fund or delivery partner is
new, this is why we get all activities considered funds and load the
buttons based on that, it's also the reason to keep the buttons outside
of the table of existing activities as these tables would not
be rendered for a new fund or delivery partner who has no activities.

This only leaves the IATI XML buttons on the organisation show page,
these have been tidied here and we have work planned to move them into
the exports journeys.