Require HistoryRecorder to define a 'trackable' association
Every historical event is associated with an activity. It is
through the lens of an activity (the 'Activity page') that
we wish to study changes.

Some historical events concern the associated activity
directly, but others concern a related item, such as a
Transaction, Budget or Forecast.

For simplicity we maintain these two relationships with two
"foreign keys":

1. `activity_id` - the `Activity` impacted

2. `trackable_id`  - the id of the object which has changed,
   often the same `Activity`, but also can be a
   `Transaction` and soon a `Budget` or a `Forecast`

 We will make the `HistoryEvent` model have a polymorphic
 association to `trackable` objects in a later commit.