Update ChangeLog
Every historical event is associated with an activity. It is
through the lens of an activity (the 'Activity page') that
we wish to study changes.

Some historical events concern the associated activity directly,
but others concern a related item, such as a `Transaction`,
`Budget` or `Forecast`.

In this PR we:

- require users of `HistoryRecorder` to supply a "trackable"
  entity in addition to an associated `Activity`

- adapt the `UpdateTransaction` service to record changes
  using the `HistoryRecorder`

- add a 'polymorphic' trackable association on
  `HistoricalEvent` to identify the entity (`Transaction` or `Activity`
  and soon `Budget` and `Forecast`) which has changed