Add a generator to generate a RODA Identifier
This uses the nanoid gem to generate a random identifier
for an activity and append it to the identifier of the parent
activity with a `-`. The identifier is always 7 characters
long and contains alphanumeric characters with all
capital letters, omitting characters like `0`, `O`, `I` and
`1`, which can easily get confused with one another.

If the activity we're generating an identifier is is also a
programme level activity, we also add the BEIS
identifier of the extending organisation to the identifier.