Allow searching by partial RODA identifier
In the last commit, we removed the
`roda_identifier_fragment` and now cannot search by
a RODA identifier's partial identifier. I've added this
commit therefore to put this functionality back, albeit
with the improvement that a user can search for a partial
RODA identifier, which does not necessarily have to
be an exact match with its fragment.

We only perform a partial match if there is no exact
match, this means users won't get a whole heap of
unnecessary results if they search for a fund or a
programme for example.