Fix: Load report activities correctly for spending report
Spending reports were throwing a:

NoMethodError (undefined method `each' for #<Activity::ProjectsForReportFinder>):
  /app/controllers/staff/spending_breakdowns_controller.rb:24

error which was causing an empty CSV file to be downloaded.

We now include an Actual "transaction" in our feature test
and verify that the downloaded file includes a row
representing the transaction we associated with our report
fixture as expected.